[Docs] Migrate using-preact example to typescript
