+++ examples/using-preact/next.config.js
@@ -1,5 +1,8 @@
 const withPreact = require('next-plugin-preact')
 
-module.exports = withPreact({
+/** @type {import('next').NextConfig} */
+const nextConfig = {
   /* regular next.js config options here */
-})
+}
+
+module.exports = withPreact(nextConfig)
