--- examples/using-preact/pages/about.js
@@ -1,3 +1,0 @@
-export default function About() {
-  return <div>About us</div>
-}
