+++ examples/using-preact/pages/about.tsx
@@ -1,0 +1,3 @@
+export default function AboutPage() {
+  return <div>About us</div>
+}
