+++ examples/using-preact/pages/index.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link'
 
-export default function Home() {
+export default function IndexPage() {
   return (
     <div>
       Hello World.{' '}
