--- examples/using-preact/pages/ssg.js
@@ -1,9 +1,0 @@
-export default function SSG({ framework }) {
-  return <div>{framework} ssg example</div>
-}
-
-export function getStaticProps() {
-  return {
-    props: { framework: 'preact' },
-  }
-}
