+++ examples/using-preact/pages/ssg.tsx
@@ -1,0 +1,13 @@
+import { InferGetStaticPropsType } from 'next'
+
+export function getStaticProps() {
+  return {
+    props: { framework: 'preact' },
+  }
+}
+
+export default function SSGPage({
+  framework,
+}: InferGetStaticPropsType<typeof getStaticProps>) {
+  return <div>{framework} ssg example</div>
+}
