--- examples/using-preact/pages/ssr.js
@@ -1,9 +1,0 @@
-export default function SSR({ framework }) {
-  return <div>{framework} ssr example</div>
-}
-
-export function getServerSideProps() {
-  return {
-    props: { framework: 'preact' },
-  }
-}
