+++ examples/using-preact/pages/ssr.tsx
@@ -1,0 +1,13 @@
+import { InferGetServerSidePropsType } from 'next'
+
+export function getServerSideProps() {
+  return {
+    props: { framework: 'preact' },
+  }
+}
+
+export default function SSRPage({
+  framework,
+}: InferGetServerSidePropsType<typeof getServerSideProps>) {
+  return <div>{framework} ssr example</div>
+}
